requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.4
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
tenacity==9.1.2
